In-Progress: fixing tests
